spi_write now returns the shifted in byte

+++ code/src/spi.h
@@ -3,7 +3,7 @@
 #include "common.h"
 
 int SPI_init(void);
-int SPI_write(char data);
+uint8_t SPI_write(char data);
 uint8_t SPI_read(void);
 void SPI_slave_select(void);
 void SPI_slave_deselect(void);
